requirements.txt
python-dotenv
SQLAlchemy>=2.0
psycopg2-binary
